[FEAT] DummyData에서 실제 DB아이템으로 이전

+++ app/src/main/java/com/example/android_accountbook_13/data/dto/History.kt
@@ -6,7 +6,7 @@ data class History(
     val methodId: Int,
     val name: String,
     val methodType: Int,
-    val money: Int,
+    val money: Long,
     val year: Int,
     val month: Int,
     val day: Int
