+++ app/src/main/java/com/example/android_accountbook_13/utils/MoneyConverter.kt
@@ -1,6 +1,6 @@
 package com.example.android_accountbook_13.utils
 
-fun moneyConverter(money: Int) : String {
+fun moneyConverter(money: Long) : String {
     val moneyString = StringBuilder(money.toString()).reverse()
     val count = (moneyString.length - 1) / 3
     for(i in 1..count) {
